# orchestra_back/src/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Orchestra API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "apiKeyAuth": {
      "type": "token",
      "in": "header",
      "name": "Authorization",
      "description": "The authorization is managed by a JWT Token provided on User login"
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/${ENTITY_ROUTE}/{organizationId}/users/{usermail}": {
      "post": {
        "tags": [
          "Organizations"
        ],
        "description": "Add an User as member of the Organization by its email",
        "parameters": [
          {
            "name": "ENTITY_ROUTE",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "usermail",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success message"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/${ENTITY_ROUTE}/{organizationId}/users/{userId}": {
      "delete": {
        "tags": [
          "Organizations"
        ],
        "description": "Removes an User from an Organization",
        "parameters": [
          {
            "name": "ENTITY_ROUTE",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "organizationId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success message"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
...
  },
...
}